Not using Properties anymore
Replaced initialization of min and max number of entries per node with
Properties by explicit constructors with such names.

Changed testing code to use the new constructors.

Added an IndexType enum for the tests so that no Strings need to be
passed around for selecting the type of index to create.